Decouple history widget.
